[hooks] Add useReceiptGeometry and refactor component
